change the location of the function I just created.
