+++ sicore/intervals.py
@@ -32,7 +32,8 @@ def _verify_and_raise(intervals):
             raise TypeError(f"Invalid data type exists in {[s, e]}")
 
         if s >= e:
-            raise ValueError(f"Inverted or no-range interval found in {[s, e]}")
+            raise ValueError(
+                f"Inverted or no-range interval found in {[s, e]}")
 
 
 def intersection(interval1, interval2, verify=False):
@@ -223,3 +224,39 @@ def poly_lt_zero(poly_or_coef, tol=1e-10):
         intervals.append([roots[-1], INF])
 
     return union_all(intervals, tol=tol)
+
+
+def polytope_to_interval(a_vec, b_vec, A=None, b=None, c=None, tol=1e-10):
+    """
+    Compute truncation intervals obtained from a polytope `{x'Ax+b'x+c<=0}` as a selection event.
+
+    Args:
+        a_vec, b_vec (array-like): Vectors which satisfy `observed_data = a_vec + b_vec * observed_test_static`
+        A (array-like, optional): `N`*`N` matrix. Set None if `A` is unused.
+            Defaults to None.
+        b (array-like, optional): `N` dimensional vector. Set None if `b` is unused.
+            Defaults to None.
+        c (float, optional): Constant. Set None if `c` is unused. Defaults to None.
+        tol (float, optional): Tolerance error parameter. Defaults to 1e-10.
+
+    Returns:
+        array-like: truncation intervals [[L1, U1], [L2, U2], ...].
+    """
+    alp = beta = gam = 0
+
+    if A is not None:
+        cA = np.dot(b_vec, A)
+        zA = np.dot(a_vec, A)
+        alp += np.dot(cA, b_vec)
+        beta += np.dot(zA, b_vec) + np.dot(cA, a_vec)
+        gam += np.dot(zA, a_vec)
+
+    if b is not None:
+        beta += np.dot(b, b_vec)
+        gam += np.dot(b, a_vec)
+
+    if c is not None:
+        gam += c
+
+    intervals = poly_lt_zero([alp, beta, gam], tol=tol)
+    return intervals
